レーザーを生成するとマップが生成できなくなってしまうバグ修正

+++ Assets/Scripts/Map/GenerateMap.cs
@@ -14,7 +14,7 @@ public class GenerateMap : MonoBehaviour
         gimmickID = new List<int>();
         bootObjectID = new List<int>();
         gimmickAssociationID = new Dictionary<int, int>();
-        rotationValue_y = new List<float>();
+        rotationValue_y = new List<int>();
     }
 
     //変数--------------------------------------------------------------------------
@@ -37,7 +37,7 @@ public class GenerateMap : MonoBehaviour
     List<int> gimmickID;                        // ギミックid
     List<int> bootObjectID;                     // ギミック起動オブジェクトid
     Dictionary<int, int> gimmickAssociationID;  // ギミックを起動オブジェクトに紐付けるためのid    
-    List<float> rotationValue_y;
+    List<int> rotationValue_y;
 
     //関数--------------------------------------------------------------------------
     // Start is called before the first frame update
@@ -201,7 +201,7 @@ public class GenerateMap : MonoBehaviour
     /// <param name="gimmickAssociationKey"> ギミックを起動オブジェクトに紐づけるためのkey </param>
     /// <param name="gimmickAssociationID"> ギミックを起動オブジェクトに紐づけるためのid </param>
     /// <param name="rotationValue_y"> ギミックのy軸回転の値 </param>
-    public void SetStageData(int layerWidth, int layerHeight, int layerNumber, List<List<List<int>>> mapData, List<int> gimmickID, List<int> bootObjectID, List<int> gimmickAssociationKey, List<int> gimmickAssociationID, List<float> rotationValue_y)
+    public void SetStageData(int layerWidth, int layerHeight, int layerNumber, List<List<List<int>>> mapData, List<int> gimmickID, List<int> bootObjectID, List<int> gimmickAssociationKey, List<int> gimmickAssociationID, List<int> rotationValue_y)
     {
         this.layerWidth = layerWidth;
         this.layerHeight = layerHeight;
